separate table components into new files
